// src/types/interfaces.ts
export interface IFormInput {
    name: string;
    email: string;
    phoneNumber: string;
    address: string;
  }
  
  export interface IFormInputVehicle {
    vehicleType: string;
    vehicleModel: string;
  }
  
  export interface IFormInputBooking {
    appointmentDate: string;
    appointmentTime: string;
  }

  